0.10

+++ HealthyMan/Models/PulseMeasurement.cs
@@ -9,7 +9,7 @@ namespace HealthyMan.Models
     public class PulseMeasurement
     {
         public int? PulseMeasurementId { get; set; }
-        public float Pulse { get; set; }
+        public int Pulse { get; set; }
         public List<float> Values { get; set; }
         public List<float> Time { get; set; }
         public Patient Patient { get; set; }
